Example1
A Dog class is made with two string and one int attributes. A new Dog object is created with a breed name, name and age value. Then getter methods are created in the Dog class to return their respective values. New strings and an int is created for breed, name and age which set using the getter methods created in class Dog. A message is then output using the new string and int variables.