Example2
Company class is created which has one String value: companyName. Then 5 company objects are created and named. A method to get the companyName was created. New strings were created and set using the getter method, and then were passed on to an array of strings that was created. I then used an enhanced for loop to iterate through the array companies, printing out their contents.